backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
Jinja2==3.1.4
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.9
psutil==5.9.8
